+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 - Get database secrets from encrypted atrribute or databag
 
-# 0.1.0
+## 0.2.0
+
+- Install git and bundler
+- Include appserver to provide Ruby
+- Create database.yml
+
+## 0.1.0
 
 - Initial release